requirements.txt
yfinance
psycopg2-binary
pymongo
fastapi
uvicorn
langchain
gradio
sqlalchemy
pandas
matplotlib
python-dotenv
langchain_community
langchain_groq

# Tests
pytest
